move dtprovision location dt->provision
