+++ backend/src/document_type/document_type.service.ts
@@ -3,18 +3,14 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { In, Repository } from 'typeorm';
 import { DocumentType } from './entities/document_type.entity';
 import { ProvisionGroup } from './entities/provision_group.entity';
-import { DocumentTypeProvision } from './entities/document_type_provision';
-// import { ProvisionService } from 'src/provision/provision.service';
 
 @Injectable()
 export class DocumentTypeService {
   constructor(
     @InjectRepository(DocumentType)
     private documentTypeRepository: Repository<DocumentType>,
-    @InjectRepository(DocumentTypeProvision)
-    private documentTypeProvisionRepository: Repository<DocumentTypeProvision>,
     @InjectRepository(ProvisionGroup)
-    private provisionGroupRepository: Repository<ProvisionGroup> // private provisionService: ProvisionService
+    private provisionGroupRepository: Repository<ProvisionGroup>
   ) {}
 
   async findById(id: number): Promise<DocumentType | undefined> {
@@ -75,126 +71,11 @@ export class DocumentTypeService {
     return this.documentTypeRepository.save(docType);
   }
 
-  /** Document Type Provision & Provision Group functions */
-
-  async getGroupMax(): Promise<any> {
-    const docTypeProvisions = await this.documentTypeProvisionRepository.find({
-      relations: ['provision_group'],
-    });
-    let provisionGroups: ProvisionGroup[] = [];
-    docTypeProvisions.forEach((provision) => {
-      provisionGroups.push(provision.provision_group);
-    });
-    provisionGroups = this.removeDuplicates(provisionGroups, 'provision_group');
-    return Array.from(provisionGroups).sort((a, b) => a.provision_group - b.provision_group);
-  }
-
   async getGroupMaxByDocTypeId(document_type_id: number): Promise<any> {
     const provisionGroups = await this.provisionGroupRepository.find({
       where: { document_type: { id: document_type_id } },
     });
     return Array.from(provisionGroups).sort((a, b) => a.provision_group - b.provision_group);
-  }
-
-  async getMandatoryProvisions(): Promise<number[]> {
-    const docTypeProvisions = await this.documentTypeProvisionRepository.find({
-      where: { type: 'M' },
-    });
-    return docTypeProvisions.map((provision) => provision.id);
-  }
-
-  async getMandatoryProvisionsByDocumentTypeId(document_type_id: number): Promise<number[]> {
-    const docTypeProvisions = await this.documentTypeProvisionRepository
-      .createQueryBuilder('provision')
-      .innerJoinAndSelect('provision.document_types', 'documentType')
-      .where('provision.type = :type', { type: 'M' })
-      .andWhere('documentType.id = :documentTypeId', { documentTypeId: document_type_id })
-      .getMany();
-
-    return docTypeProvisions.map((provision) => provision.id);
-  }
-
-  async getManageDocTypeProvisions(document_type_id: number) {
-    // type used by the frontend for displaying data in a table
-    type ManageDocTypeProvision = {
-      id: number;
-      type: string;
-      provision_name: string;
-      free_text: string;
-      help_text: string;
-      category: string;
-      active_flag: boolean;
-      sequence_value: number;
-      associated: boolean; // is the provision currently associated with the doc type
-      provision_group: ProvisionGroup | null;
-    };
-
-    // get all docTypeProvisions with provision_group relation
-    const docTypeProvisions = await this.documentTypeProvisionRepository.find({
-      relations: ['provision_group', 'provision'],
-    });
-    console.log(docTypeProvisions);
-    // get provisions which are associated with the document type
-    const associatedProvisions = await this.documentTypeProvisionRepository.find({
-      where: { document_type: { id: document_type_id } },
-    });
-    console.log(associatedProvisions);
-    // map them to a more readable format
-    const associatedProvisionIds = associatedProvisions.map((raw) => raw.id);
-    // map 'docTypeProvisions' to include a variable called 'associated' which is true or false
-    const managedProvisions: ManageDocTypeProvision[] = docTypeProvisions.map((docTypeProvision) => ({
-      id: docTypeProvision.id,
-      type: docTypeProvision.type,
-      provision_name: docTypeProvision.provision.provision_name,
-      free_text: docTypeProvision.provision.free_text,
-      help_text: docTypeProvision.provision.help_text,
-      category: docTypeProvision.provision.category,
-      active_flag: docTypeProvision.provision.active_flag,
-      sequence_value: docTypeProvision.sequence_value,
-      associated: associatedProvisionIds.includes(docTypeProvision.id),
-      provision_group: docTypeProvision.provision_group,
-    }));
-    return managedProvisions;
-  }
-
-  async associateDocType(provision_id: number, document_type_id: number) {
-    try {
-      console.log(`associating provision id ${provision_id} to doc type ${document_type_id}`);
-      const documentType: DocumentType = await this.findById(document_type_id);
-      const provision: DocumentTypeProvision = await this.documentTypeProvisionRepository.findOneBy({
-        id: provision_id,
-      });
-      // make sure they aren't already related
-      const existingProvision = documentType.document_type_provisions.find((p) => p.id === provision.id);
-      if (documentType && provision && !existingProvision) {
-        documentType.document_type_provisions.push(provision);
-        await this.save(documentType);
-      } else {
-        throw new Error(`Failed to associate Provision with Document Type`);
-      }
-    } catch (err) {
-      console.log(err);
-      throw err;
-    }
-  }
-
-  async disassociateDocType(provision_id: number, document_type_id: number) {
-    try {
-      console.log(`disassociating provision id ${provision_id} from doc type ${document_type_id}`);
-      const documentType: DocumentType = await this.findById(document_type_id);
-      const provisionIndex = documentType.document_type_provisions.findIndex((p) => {
-        return p.id === provision_id;
-      });
-      if (documentType && provisionIndex > -1) {
-        documentType.document_type_provisions.splice(provisionIndex, 1);
-        await this.save(documentType);
-      } else {
-        throw new Error(`Failed to disassociate Provision and Document Type`);
-      }
-    } catch (err) {
-      console.log(err);
-      throw err;
-    }
   }
 
   async updateGroupMaximums(provision_group: number, max: number, provision_group_text: string) {
