+++ backend/src/provision/entities/document_type_provision.ts
@@ -1,13 +1,16 @@
 import { DocumentDataProvision } from 'src/document_data/entities/document_data_provision.entity';
 import { Column, Entity, ManyToOne, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { ProvisionGroup } from './provision_group.entity';
-import { Provision } from '../../provision/entities/provision.entity';
+import { ProvisionGroup } from '../../document_type/entities/provision_group.entity';
+import { Provision } from './provision.entity';
 import { DocumentType } from 'src/document_type/entities/document_type.entity';
 
 @Entity()
 export class DocumentTypeProvision {
   @PrimaryGeneratedColumn()
   id: number;
+
+  @Column()
+  associated: boolean;
 
   @Column({ nullable: true })
   sequence_value: number;
@@ -28,4 +31,10 @@ export class DocumentTypeProvision {
     cascade: true,
   })
   document_data_provisions: DocumentDataProvision[];
+
+  constructor(associated?: boolean, sequence_value?: number, type?: string) {
+    this.associated = associated || false;
+    this.sequence_value = sequence_value || 1;
+    this.type = type || 'O';
+  }
 }
