+++ backend/src/report/report.service.ts
@@ -8,7 +8,6 @@ import { GL_REPORT_TYPE, LUR_REPORT_TYPE, numberWords, sectionTitles } from 'uti
 import { ProvisionJSON, VariableJSON } from 'utils/types';
 import { convertToSpecialCamelCase, formatMoney, grazingLeaseVariables, nfrAddressBuilder } from 'utils/util';
 import { DocumentDataService } from 'src/document_data/document_data.service';
-import { DocumentTemplate } from 'src/document_template/entities/document_template.entity';
 import { DocumentDataLogService } from 'src/document_data_log/document_data_log.service';
 import { DocumentTypeService } from 'src/document_type/document_type.service';
 import { Provision } from 'src/provision/entities/provision.entity';
@@ -774,12 +773,8 @@ export class ReportService {
     }
   }
 
-  getMandatoryProvisions() {
-    return this.documentTypeService.getMandatoryProvisions();
-  }
-
   getMandatoryProvisionsByDocumentTypeId(document_type_id: number) {
-    return this.documentTypeService.getMandatoryProvisionsByDocumentTypeId(document_type_id);
+    return this.provisionService.getMandatoryProvisionsByDocumentTypeId(document_type_id);
   }
 
   async getDocumentData() {
@@ -869,7 +864,7 @@ export class ReportService {
   }
 
   async getEnabledProvisionsByDocTypeId(document_type_id: number) {
-    return this.documentTypeService.getMandatoryProvisionsByDocumentTypeId(document_type_id);
+    return this.provisionService.getMandatoryProvisionsByDocumentTypeId(document_type_id);
   }
 
   getEnabledProvisionsByDocTypeIdDtid(document_type_id: number, dtid: number) {
